Revert "Convert remaining objective c code to swift"
This reverts commit 703f26fdfe2eea3054a1e0157d828a8691e7a4b1.

+++ TradeItIosTicketSDK2/TradeItFxOrderCapabilities.swift
@@ -1,3 +1,3 @@
 class TradeItFxOrderCapabilities: TradeItInstrumentOrderCapabilities {
-    var leverageOptions: [Int]?
+    var leverageOptions: [Double]?
 }
